BACKLOG-11925: move pickerDialog from design system to the picker selectorType
